// packages/my-lib/src/swap/types.ts
export type SwapAmountInputReact = {
  /** Optional className override for top div element */
  className?: string;
  /** The debounce delay in milliseconds */
  delayMs?: number;
  /** Descriptive label for the input field */
  label: string;
  /** Swappable tokens */
  swappableTokens?: Token[];
  /** Selected token */
  token?: Token;
  /** Identifies if component is for toToken or fromToken */
  type: 'to' | 'from';
};